requirements.txt
appdirs==1.4.4
asgiref==3.3.1
boto3==1.16.13
botocore==1.19.13
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
fs==2.4.11
fs-s3fs==1.1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.7
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11